Minor javadoc fix
git-svn-id: svn://svn.hafro.is/jminor/trunk@8456 51816742-9f83-410d-85ed-e7bcf7042953

+++ common/core/src/test/java/org/jminor/common/DateUtilTest.java
@@ -80,7 +80,7 @@ public class DateUtilTest {
     assertEquals(0, calendar.get(Calendar.MINUTE));
     assertEquals(42, calendar.get(Calendar.SECOND));
     assertEquals(0, calendar.get(Calendar.MILLISECOND));
-    DateUtil.floorFields(calendar, null);
+    DateUtil.floorFields(calendar, (Integer) null);
   }
 
   @Test(expected = IllegalArgumentException.class)
